fix: menu component
