+++ packages/core/src/react/blocks/ButtonBlock.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Element, Text } from 'html-react-parser';
 import { isBlock } from '../../dom';
 import { IBlock } from '../components';
