+++ packages/core/src/react/blocks/HeadingBlock.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { isBlockByName } from '../../dom';
 import { DEFAULT_BLOCK_ELEMENT } from '../../dom/parseBlockAttributes';
 import { IBlock } from '../components';
