+++ packages/core/src/react/components/Menu.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 /* eslint-disable react/no-unused-prop-types */
 /* eslint-disable @typescript-eslint/no-use-before-define, react/require-default-props */
 import { PropsWithChildren } from 'react';
