+++ packages/core/src/react/hooks/useFetchMenu.ts
@@ -1,42 +1,10 @@
-import { AppEntity, FetchResponse, MenuItemEntity } from '../../data';
+import { AppEntity, FetchResponse, flatToHierarchical, MenuItemEntity } from '../../data';
 import { FetchHookOptions, HookResponse } from './types';
 import { useFetchAppSettings } from './useFetchAppSettings';
 import { isProxy, makeErrorCatchProxy } from './util';
 
 export interface useMenuResponse extends HookResponse {
 	data: MenuItemEntity[];
-}
-
-function flatToHierarchical(flat: MenuItemEntity[]): MenuItemEntity[] {
-	const roots: MenuItemEntity[] = [];
-
-	const all: Record<number, MenuItemEntity> = {};
-	flat.forEach((item, index) => {
-		all[item.ID] = { ...item, children: [], order: index };
-	});
-
-	Object.keys(all).forEach((key) => {
-		const id = Number(key);
-		const item = all[id];
-		const parentId = Number(item.menu_item_parent);
-
-		if (parentId === 0) {
-			roots.push(item);
-		} else if (item.menu_item_parent in all) {
-			const p = all[item.menu_item_parent];
-			if (!('children' in p)) {
-				p.children = [];
-			}
-			p.children.push(item);
-		}
-	});
-
-	roots.sort((a, b) => a.order - b.order);
-	roots.forEach((root) => {
-		root?.children?.sort((a, b) => a.order - b.order);
-	});
-
-	return roots;
 }
 
 /**
