+++ projects/wp-nextjs-app/src/app/(single)/[...path]/page.tsx
@@ -4,6 +4,9 @@ import { HeadstartWPRoute, queryPost } from '@headstartwp/next/app';
 const Single = async ({ params }: HeadstartWPRoute) => {
 	const { data } = await queryPost({
 		routeParams: params,
+		params: {
+			postType: ['post', 'page'],
+		},
 		options: {
 			headers: {
 				cache: 'force-cache',
